# openapi.yaml
openapi: 3.1.0
info:
  title: Dark Factory API
  version: 0.1.0
  description: API-first coordination runtime for humans and agents.
servers:
  - url: http://localhost:3000/api/v1
    description: Local
...
paths:
  /agents/register:
    post:
      tags: [Agents]
      summary: Register an agent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegisterRequest'
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
...
  /artifacts:
    post:
      tags: [Artifacts]
      summary: Attach artifact reference
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ArtifactCreateRequest'
      responses:
        '201':
          description: Artifact created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artifact'
...
components:
...
  schemas:
    Agent:
      type: object
      properties:
        id: { type: string, format: uuid }
        agent_key: { type: string }
        name: { type: string }
        description: { type: string }
        type: { type: string }
        capabilities:
          type: array
          items: { type: string }
        status:
          type: string
          enum: [active, inactive]
      required: [id, agent_key, name, type, status]
...
    Artifact:
      type: object
      properties:
        id: { type: string, format: uuid }
        task_id: { type: string, format: uuid }
        workflow_run_id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        uri: { type: string }
        metadata_json: { type: object, additionalProperties: true }
        created_by_actor_id: { type: string, format: uuid }
        approved_status:
          type: string
          enum: [unreviewed, approved, rejected]
      required: [id, kind, title, uri, approved_status]
...
    AgentRegisterRequest:
      type: object
      properties:
        agent_key: { type: string }
        name: { type: string }
        description: { type: string }
        type: { type: string }
        capabilities:
          type: array
          items: { type: string }
      required: [agent_key, name]
...
    ArtifactCreateRequest:
      type: object
      properties:
        task_id: { type: string, format: uuid }
        workflow_run_id: { type: string, format: uuid }
        kind: { type: string }
        title: { type: string }
        uri: { type: string }
        metadata_json: { type: object, additionalProperties: true }
        created_by_actor_id: { type: string, format: uuid }
      required: [kind, title, uri]
...